moved middleware to external file
