+++ middleware.js
@@ -1,3 +1,8 @@
+// VALIDATION
+const ExpressError = require('./utils/ExpressError');
+const { playgroundSchema, reviewSchema } = require('./schemas.js');
+
+// CHECK IF USER IS LOGGED IN
 module.exports.isLoggedIn = (req, res, next) => {
   console.log('REQ.USER: ', req.user);
   if (!req.isAuthenticated()) {
@@ -5,5 +10,41 @@ module.exports.isLoggedIn = (req, res, next) => {
     req.flash('error', 'You must log in to use this feature.');
     return res.redirect('/login');
   };
+  next();
+};
+
+// ERROR HANDLING
+module.exports.validatePlayground = (req, res, next) => {
+  const { error } = playgroundSchema.validate(req.body);
+  if (error) {
+    const msg = error.details.map(item => item.message).join(',');
+    throw new ExpressError(msg, 400);
+  } else {
+    next();
+  }
+};
+
+module.exports.validateReview = (req, res, next) => {
+  const { error } = reviewSchema.validate(req.body);
+  if (error) {
+    const msg = error.details.map(item => item.message).join(',');
+    throw new ExpressError(msg, 400);
+  } else {
+    next();
+  };
+};
+
+// CHECK IF CURRENT USER IS AUTHOR
+module.exports.isAuthor = async (req, res, next) => {
+  const { id } = req.params;
+  const playground = await Playground.findById(id);
+  if (!playground) {
+    req.flash('error', 'Sorry, this playground could not be found.');
+    return res.redirect('/playgrounds');
+  }
+  if (!playground.author.equals(req.user._id)) {
+    req.flash('error', 'You do not have permission to edit this playground.');
+    return res.redirect(`/playgrounds/${id}`);
+  }
   next();
 };
