+++ routes/playgrounds.js
@@ -1,25 +1,10 @@
 const express = require('express'),
   catchAsync = require('../utils/catchAsync'),
-  ExpressError = require('../utils/ExpressError'),
   Playground = require('../models/Playground'),
   router = express.Router();
 
-// VALIDATION
-const { playgroundSchema } = require('../schemas.js');
-
-// AUTHORIZATION
-const { isLoggedIn } = require('../middleware');
-
-// ERROR HANDLING
-const validatePlayground = (req, res, next) => {
-  const { error } = playgroundSchema.validate(req.body);
-  if (error) {
-    const msg = error.details.map(item => item.message).join(',');
-    throw new ExpressError(msg, 400);
-  } else {
-    next();
-  }
-};
+// MIDDLEWARE
+const { isLoggedIn, isAuthor, validatePlayground } = require('../middleware');
 
 // READ
 router.get(
@@ -51,16 +36,8 @@ router.get(
 router.get(
   '/:id/edit',
   isLoggedIn,
+  isAuthor,
   catchAsync(async (req, res) => {
-    const playground = await Playground.findById(req.params.id);
-    if (!playground) {
-      req.flash('error', 'Sorry, this playground could not be found.');
-      return res.redirect('/playgrounds');
-    };
-    if (!playground.author.equals(req.user._id)) {
-      req.flash('error', 'You do not have permission to edit this playground.');
-      return res.redirect(`/playgrounds/${req.params.id}`);
-    };
     const editedPlayground = await Playground.findById(req.params.id);
     res.render('playgrounds/edit', { playground });
   })
@@ -84,15 +61,13 @@ router.post(
 router.put(
   '/:id',
   isLoggedIn,
+  isAuthor,
   validatePlayground,
   catchAsync(async (req, res) => {
     const { id } = req.params;
-    const playground = await Playground.findById(id);
-    if (!playground.author.equals(req.user._id)) {
-      req.flash('error', 'You do not have permission to edit this playground.');
-      return res.redirect(`/playgrounds/${id}`);
-    };
-    const editedPlayground = await Playground.findByIdAndUpdate(id, { ...req.body.playground });
+    const editedPlayground = await Playground.findByIdAndUpdate(id, {
+      ...req.body.playground,
+    });
     req.flash('success', 'Successfully updated this playground!');
     res.redirect(`/playgrounds/${id}`);
   })
@@ -101,17 +76,10 @@ router.put(
 // DELETE
 router.delete(
   '/:id',
+  isLoggedIn,
+  isAuthor,
   catchAsync(async (req, res) => {
     const { id } = req.params;
-    const playground = await Playground.findById(id);
-    if (!playground) {
-      req.flash('error', 'Sorry, this playground could not be found.');
-      return res.redirect('/playgrounds');
-    };
-    if (!playground.author.equals(req.user._id)) {
-      req.flash('error', 'You do not have permission to delete this playground.');
-      return res.redirect(`/playgrounds/${id}`);
-    };
     const deletedPlayground = await Playground.findByIdAndDelete(id);
     req.flash('success', 'The playground has been removed.');
     res.redirect('/playgrounds');
